Add a `PThreadPoolExecutor`
# Motivation

In a previous PR, I landed support for a simple `PThreadExecutor` which is backed by a single pthread and epoll/kqueue.

# Modification

This PR introduces a new pooled executor of `PThreadExecutor`s which is useful in context where one wants to parallize work across multiple threads.

# Result

We now can create a pool of pthread executors.