Improve performance of executors and add `PlatformExecutorFactory`.
# Motivation

We want our new executors to be as fast as the current ones. This means we need to ensure that there are no unnecessary re-enqueues or context switches.

# Modifications

This PR drops any conformance to `Serial` or `Task` executor on types that already have the other. This is due to a performance foot-gun in the runtime which means no type should conform to both at once. This PR also provides the `PlatfromExecutorFactory`.

Lastly, this PR also adds an example target showing how to use some of the new executors.

# Result
We now can use our new executors as defaults and they have matching performance with the current ones