Add support for job priorities in the PThreadExecutor
# Motivation

We want to support priorities in our executor so that they handle different task priorities correctly.

# Modifications

This PR adds support for priorities to the PThreadExecutor by using the existing PriorityQueue. It also allows us to drop the dependency on swift-collections. Lastly, this PR enables running the PThread executors through the existing executor test fixtures.

# Result

We properly support priorities now.